Move tests to test_cuda_array_interface
